requirements.txt
h5py
hdijupyterutils
html5lib
ipykernel
ipython
ipython-genutils
jsonschema
jupyter-client
jupyter-core
nltk
notebook
numpy
pandas
prompt-toolkit
protobuf
pycparser
pyparsing
python-dateutil
PyYAML
scikit-learn
scipy
torch
torchvision
tqdm


